+++ CHANGELOG.md
@@ -26,6 +26,7 @@ XQuery:
 XProc:
 
 1. Recognise xpl files as XProc XML files.
+1. Add the XProc 1.0 RNG schema file.
 
 XRay Unit Tests: